Windows Optimization * Path sanitizing
1. Handle ':' in Windows drive names, do not sanitze ':' under Windows
2. This is just a quick and dirty hack. The ':' charcter should not be
sanitized in Windows drive names only. If there are more ':', these
should be sanitized jsut like in the standard case

* FF preferences

1. Add ability to find and load FF preferences
2. Used mainly for proxying functionality

* Deployment configuration

1. On Windows File.renameTo() returns false if the destination already
exists. So backup will fail if a backupfile already exists
2. Handle backup of deployment configuration on Windows

* More Windows fiendly log file names

* Resource locking

1. Keep track of the FileDescriptor instance bound to the locked
resource
2. Use the tracked FileDescriptor instance for read operations on locked
resources

* Resource locking

1. Keep track of the FileDescriptor instance bound to the locked
resource
2. Use the tracked FileDescriptor instance for read operations on locked
resources

* JNLP cache deletion

1. Windows specific handling to make possible JNLP cache deletion
2. Trucates the file which tracks the cached resources

* JNLP cache deletion

1. Windows specific handling to make possible JNLP cache deletion
2. Trucates the file which tracks the cached resources

* Policy editor

1. Make possible java policy editing on Windows
2. Use FileLock insted of LockedFile

* Resource fetching

1. Ignore FileNotFoundException (caused by HTTP 404) upon resource
pre-fetching

+++ netx/net/sourceforge/jnlp/util/FileUtils.java
@@ -99,6 +99,12 @@ public final class FileUtils {
         for (int i = 0; i < INVALID_CHARS.length; i++) {
             if (INVALID_CHARS[i] != File.separatorChar) {
                 if (-1 != path.indexOf(INVALID_CHARS[i])) {
+                	if (JNLPRuntime.isWindows() && INVALID_CHARS[i] == ':') {
+            			// Do not sanitize a colon (':') on a Windows OS
+                		// TODO : This a quick and dirty hack. 
+                		// Find a better way to handle Windows drive names
+            			continue;
+                	}
                     path = path.replace(INVALID_CHARS[i], substitute);
                 }
             }
@@ -122,6 +128,12 @@ public final class FileUtils {
 
         for (int i = 0; i < INVALID_CHARS.length; i++) {
             if (-1 != filename.indexOf(INVALID_CHARS[i])) {
+            	if (JNLPRuntime.isWindows() && INVALID_CHARS[i] == ':') {
+        			// Do not sanitize a colon (':') on a Windows OS
+            		// TODO : This a quick and dirty hack. 
+            		// Find a better way to handle Windows drive names
+        			continue;
+            	}
                 filename = filename.replace(INVALID_CHARS[i], substitute);
             }
         }
@@ -258,6 +270,12 @@ public final class FileUtils {
             // rename this file. Unless the file is moved/renamed, any program that
             // opened the file right after it was created might still be able to
             // read the data.
+            // On Windows File.renameTo returns false if the "rename" destination already exists
+            if (file.exists() && !file.delete()) {
+            	OutputController.getLogger().log(OutputController.Level.ERROR_ALL, R("RCantRename", tempFile, file));
+            	return;
+            }
+            
             if (!tempFile.renameTo(file)) {
                 OutputController.getLogger().log(OutputController.Level.ERROR_ALL, R("RCantRename", tempFile, file));
             }
